+++ platform/src/main/java/org/stellar/anchor/platform/data/JdbcSep31Refunds.java
@@ -3,7 +3,7 @@ package org.stellar.anchor.platform.data;
 import com.google.gson.annotations.SerializedName;
 import java.util.List;
 import lombok.Data;
-import org.stellar.anchor.model.Sep31Transaction;
+import org.stellar.anchor.sep31.Sep31Transaction;
 
 @Data
 public class JdbcSep31Refunds implements Sep31Transaction.Refunds {
