+++ platform/src/main/java/org/stellar/anchor/platform/data/JdbcSep31Transaction.java
@@ -9,7 +9,7 @@ import java.util.Set;
 import javax.persistence.*;
 import lombok.Data;
 import org.stellar.anchor.api.sep.AssetInfo;
-import org.stellar.anchor.model.Sep31Transaction;
+import org.stellar.anchor.sep31.Sep31Transaction;
 import org.stellar.anchor.util.GsonUtils;
 
 @Data
