+++ platform/src/main/java/org/stellar/anchor/platform/data/JdbcSep31TransactionRepo.java
@@ -7,7 +7,7 @@ import lombok.NonNull;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
-import org.stellar.anchor.model.Sep31Transaction;
+import org.stellar.anchor.sep31.Sep31Transaction;
 
 public interface JdbcSep31TransactionRepo extends CrudRepository<JdbcSep31Transaction, String> {
   Optional<JdbcSep31Transaction> findById(@NonNull String id);
