+++ platform/src/main/java/org/stellar/anchor/platform/data/JdbcSep38QuoteStore.java
@@ -2,7 +2,7 @@ package org.stellar.anchor.platform.data;
 
 import lombok.NonNull;
 import org.stellar.anchor.api.exception.SepException;
-import org.stellar.anchor.model.Sep38Quote;
+import org.stellar.anchor.sep38.Sep38Quote;
 import org.stellar.anchor.sep38.Sep38QuoteStore;
 
 public class JdbcSep38QuoteStore implements Sep38QuoteStore {
