+++ platform/src/main/java/org/stellar/anchor/platform/service/TransactionService.java
@@ -24,8 +24,8 @@ import org.stellar.anchor.api.sep.AssetInfo;
 import org.stellar.anchor.api.shared.Amount;
 import org.stellar.anchor.api.shared.Transaction;
 import org.stellar.anchor.asset.AssetService;
-import org.stellar.anchor.model.Sep31Transaction;
-import org.stellar.anchor.model.Sep38Quote;
+import org.stellar.anchor.sep31.Sep31Transaction;
+import org.stellar.anchor.sep38.Sep38Quote;
 import org.stellar.anchor.platform.data.JdbcSep31Transaction;
 import org.stellar.anchor.platform.data.JdbcSep31TransactionStore;
 import org.stellar.anchor.sep38.Sep38QuoteStore;
