+++ sep/src/main/java/org/stellar/anchor/sep24/Sep24TransactionBuilder.java
@@ -1,6 +1,4 @@
-package org.stellar.anchor.model;
-
-import org.stellar.anchor.sep24.Sep24TransactionStore;
+package org.stellar.anchor.sep24;
 
 @SuppressWarnings("unused")
 public class Sep24TransactionBuilder {
