+++ sep/src/main/java/org/stellar/anchor/sep31/Sep31TransactionStore.java
@@ -5,7 +5,6 @@ import java.util.List;
 import lombok.NonNull;
 import org.stellar.anchor.api.exception.AnchorException;
 import org.stellar.anchor.api.exception.SepException;
-import org.stellar.anchor.model.Sep31Transaction;
 
 public interface Sep31TransactionStore {
   Sep31Transaction newTransaction();
