+++ sep/src/test/java/org/stellar/anchor/sep38/PojoSep38Quote.java
@@ -2,7 +2,6 @@ package org.stellar.anchor.sep38;
 
 import java.time.Instant;
 import lombok.Data;
-import org.stellar.anchor.model.Sep38Quote;
 
 @Data
 public class PojoSep38Quote implements Sep38Quote {
